refactor(plugins): Move ftplugin/syntax/indent to start
No benefit loading them later since they are loading by rust based on
filetype anyway, plus this solves on html issue. In ftplugin remove
did_ftplugin since these are not in the after directory and should not
use this or else they will not fire since other ftpluins that come
before it will set that variable.

Remove old plugins either not updated or not in use by me.

closes #7